Make tests compatible with changes
Removed AnomalyDetectionTests, changed imports so they work.

Added an instruction in Makefile
