+++ tests/fixtures/anomaly_detection_sns.json
