+++ tests/fixtures/query_write_cloudtrail.json
